iff2lvl: port to 64-bit Linux; build OAD pipeline; verify primitives.lev
- scalar.hpi (linux): replace x86 inline asm with portable C using long
  (64-bit on Linux x86-64) for Invert, operator*=, /=, MulDiv, Multiply64,
  AddCarry64, SubCarry64, ShiftRight*, JoinHiLo, FixedMac3
- scalar.hp: long _value → int32 _value (fixed-point value must be 32-bit)
- assert.hp: fix _sys_assert(int(exp),...) pointer-to-int cast on 64-bit
- level.hp, level4.cc: long* → int32* in contextAnimArray signatures
- point6.cc: remove invalid #pragma message inside expression (Watcom-ism)
- GNUMakefile.linux: add xlib-stubs path so -lXext/-lXmu/-lXi link on Ubuntu
- oas/iff_with_kinds.s: wrapper that includes objects.p before iff.s so
  _KIND enum values resolve when building OAD .iff files
- oas/objects.lc, objects.p: regenerated from objects.mac (proper names,
  no Gold; objects.lc TODO: generate as part of build)
- oas/assets.txt: asset extension table needed by iff2lvl
- oas/gold.oas: stub OAD so objects.lc entries compile (TODO: real file)
- docs: pipeline investigation doc and showas screenshot width fix

Verified: primitives.lev → iffcomp → primitives.lev.bin → iff2lvl →
primitives.lvl (7532 bytes, no errors)

+++ wftools/iff2lvl/level4.cc
@@ -661,7 +661,7 @@ QLevel::AddXDataToCommonData(void)
 //============================================================================
 
 int32
-QLevel::ParseTextFileIntoObjectList(char* textBuffer, int32 sourceSize, long* objectReferenceArray,int32 maxEntries)
+QLevel::ParseTextFileIntoObjectList(char* textBuffer, int32 sourceSize, int32* objectReferenceArray,int32 maxEntries)
 {
 #if 0
 	cout << "ParseTextFileIntoObjectList: sourceSize = " << sourceSize << endl;
@@ -702,7 +702,7 @@ QLevel::ParseTextFileIntoObjectList(char* textBuffer, int32 sourceSize, long* ob
 int32
 QLevel::ParseTextFileIntoContextualAnimationList(char* textBuffer,
 												 int32 sourceSize,
-												 long contextAnimArray[][2],
+												 int32 contextAnimArray[][2],
 												 int32 maxEntries,
 												 int32 roomNum)
 {
